Restructure folder structure
