+++ communication/stream_service.py
