+++ neuro_processing/primary_dataloops/start.py
@@ -1,7 +1,7 @@
 """ Import fix - check README for documentation """ 
 import os,sys,inspect 
 __currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-sys.path.insert(0, __currentdir[0:__currentdir.find("software/src")+len("software/src")])
+sys.path.insert(0, __currentdir[0:__currentdir.find("CREPE")+len("CREPE")])
 """ End import fix """
 
 # This file will keep track of which start the primary dataloop.
